refactor(storage_impl): Move Conversion impl for payment_attempt to storage_impl crate
