+++ crates/storage_impl/src/mock_db/payment_attempt.rs
@@ -3,15 +3,13 @@ use common_utils::errors::CustomResult;
 use common_utils::id_type;
 use diesel_models::enums as storage_enums;
 use error_stack::ResultExt;
-#[cfg(feature = "v1")]
-use hyperswitch_domain_models::behaviour::Conversion;
 use hyperswitch_domain_models::{
     merchant_key_store::MerchantKeyStore,
     payments::payment_attempt::{PaymentAttempt, PaymentAttemptInterface, PaymentAttemptUpdate},
 };
 
 use super::MockDb;
-use crate::errors::StorageError;
+use crate::{behaviour::Conversion, errors::StorageError};
 
 #[async_trait::async_trait]
 impl PaymentAttemptInterface for MockDb {
